# src: add a method to get string encoding info