--- a/doc/api/v8.md
+++ b/doc/api/v8.md
@@ -1313,7 +1313,7 @@ added: REPLACEME
 * `content` {string}
 * Returns: {boolean}
 
-V8 only supports `Latin-1/ISO-8859-1` and `UTF16` as the underlying representations.
+V8 only supports `Latin-1/ISO-8859-1` and `UTF16` as the underlying representation of a string.
 If the `content` uses `Latin-1/ISO-8859-1` as the underlying representation, this function will return true;
 otherwise, it returns false.
 
@@ -1322,7 +1322,7 @@ Sometimes a `Latin-1` string may also be represented as `UTF16`.
 
 ```js
 const { isStringOneByteRepresentation } = require('node:v8');
-const assert = require('node:assert');
+const { strictEqual } = require('node:assert');
 
 const Encoding = {
   latin1: 1,
@@ -1409,8 +1409,8 @@ const data = serializer.finish();
 
 // Read strings from the buffer.
 const deserializer = new Deserializer(data);
-assert(deserializer.readString() === 'hello');
-assert(deserializer.readString() === '你好');
+strictEqual(deserializer.readString(), 'hello');
+strictEqual(deserializer.readString(), '你好');
 ```
 
 [HTML structured clone algorithm]: https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API/Structured_clone_algorithm