--- a/doc/api/v8.md
+++ b/doc/api/v8.md
@@ -1322,95 +1322,25 @@ Sometimes a `Latin-1` string may also be represented as `UTF16`.
 
 ```js
 const { isStringOneByteRepresentation } = require('node:v8');
-const { strictEqual } = require('node:assert');
 
 const Encoding = {
   latin1: 1,
   utf16le: 2,
 };
-
-/**
- * Read strings from the buffer.
- *
- * Note that this example ignores flag checks and boundary checks.
- */
-class Deserializer {
-  buffer;
-  cursor;
-
-  constructor(buffer) {
-    this.buffer = buffer;
-    this.cursor = 0;
-  }
-
-  readString() {
-    const encoding = this.buffer.readUint8(this.cursor++);
-    const length = this.buffer.readUint32LE(this.cursor);
-    this.cursor += 4;
-    if (encoding === Encoding.latin1) {
-      const result = this.buffer.toString('latin1', this.cursor, this.cursor + length);
-      this.cursor += length;
-      return result;
-    }
-    const result = this.buffer.toString('utf16le', this.cursor, this.cursor + length);
-    this.cursor += length;
-    return result;
+const buffer = Buffer.alloc(100);
+function writeString(input) {
+  if (isStringOneByteRepresentation(input)) {
+    buffer.writeUint8(Encoding.latin1);
+    buffer.writeUint32LE(input.length, 1);
+    buffer.write(input, 5, 'latin1');
+  } else {
+    buffer.writeUint8(Encoding.utf16le);
+    buffer.writeUint32LE(input.length * 2, 1);
+    buffer.write(input, 5, 'utf16le');
   }
 }
-
-/**
- * By means of the `isStringOneByteRepresentation` function,
- * we can write strings into the buffer with high performance,
- * and it only takes the time of a memcopy.
- *
- * Note that this example ignores boundary checks.
- */
-class Serializer {
-  buffer;
-  cursor;
-
-  constructor() {
-    this.buffer = Buffer.alloc(100);
-    this.cursor = 0;
-  }
-
-  /**
-   * step1: Write the encoding flag to the first byte.
-   * step2: Write the byte length of the string to the next four bytes.
-   * step3: Write the string to the buffer.
-   * @param {string} input
-   */
-  writeString(input) {
-    if (isStringOneByteRepresentation(input)) {
-      this.buffer.writeUint8(Encoding.latin1, this.cursor++);
-      this.buffer.writeUint32LE(input.length, this.cursor);
-      this.cursor += 4;
-      this.buffer.write(input, this.cursor, 'latin1');
-      this.cursor += input.length;
-    } else {
-      this.buffer.writeUint8(Encoding.utf16le, this.cursor++);
-      this.buffer.writeUint32LE(input.length * 2, this.cursor);
-      this.cursor += 4;
-      this.buffer.write(input, this.cursor, 'utf16le');
-      this.cursor += input.length * 2;
-    }
-  }
-
-  finish() {
-    return this.buffer.subarray(0, this.cursor);
-  }
-}
-
-// Write strings to the buffer.
-const serializer = new Serializer();
-serializer.writeString('hello');
-serializer.writeString('你好');
-const data = serializer.finish();
-
-// Read strings from the buffer.
-const deserializer = new Deserializer(data);
-strictEqual(deserializer.readString(), 'hello');
-strictEqual(deserializer.readString(), '你好');
+writeString('hello');
+writeString('你好');
 ```
 
 [HTML structured clone algorithm]: https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API/Structured_clone_algorithm