--- a/lib/v8.js
+++ b/lib/v8.js
@@ -104,6 +104,7 @@ const binding = internalBinding('v8');
 const {
   cachedDataVersionTag,
   setFlagsFromString: _setFlagsFromString,
+  isStringOneByteRepresentation: _isStringOneByteRepresentation,
   updateHeapStatisticsBuffer,
   updateHeapSpaceStatisticsBuffer,
   updateHeapCodeStatisticsBuffer,
@@ -154,6 +155,17 @@ function setFlagsFromString(flags) {
   validateString(flags, 'flags');
   _setFlagsFromString(flags);
 }
+
+/**
+ * Return whether this string uses one byte as underlying representation or not.
+ * @param {string} content
+ * @returns {boolean}
+ */
+function isStringOneByteRepresentation(content) {
+  validateString(content, 'content');
+  return _isStringOneByteRepresentation(content);
+}
+
 
 /**
  * Gets the current V8 heap statistics.
@@ -439,4 +451,5 @@ module.exports = {
   startupSnapshot,
   setHeapSnapshotNearHeapLimit,
   GCProfiler,
+  isStringOneByteRepresentation,
 };