--- a/test/parallel/test-v8-string-is-one-byte-representation.js
+++ b/test/parallel/test-v8-string-is-one-byte-representation.js
@@ -1,4 +1,3 @@
-// Flags: --expose-internals
 'use strict';
 require('../common');
 const assert = require('assert');