util:  Add a method to get the encoding information of a string.

Currently,  we should check the encoding of the string before  we use  `buffer.write("string", 'latin1')` and `buffer.write("string", 'utf16le')` to write to the buffer.
This PR adds a method to return the encoding information from V8. 

Closes: #56090