@theweipeng 2024-12-18
> Actually, I see that @joyeecheung already left a number of helpful comments on the original ticket. Moving this to the `v8` API makes a bit more sense, because it really is engine-specific, and she suggests a better name for it that reflects the subtleties involved here (`isOneByteRepresentation`), plus points out that `utf16le` isn't accurate because we'd need to take platform endianness into account.
> 
> So – Joyee made a lot of good suggestions here already, and you'll probably just want to incorporate them.

Understood, I'll correct the API placement. Thanks for pointing it out.