@theweipeng 2024-12-31
> There is still no clear use case for this. The OP mentions `ucs2Write` and `latin1WriteStatic` but these are internal methods so cannot be used as arguments to add this method publicly.

Sorry, I had a problem with the expression in the OP. I didn't think to use these private APIs; I used `buffer.write` in the documentation. I have corrected the OP.