@theweipeng 2024-12-31
> Can you please answer @addaleax's question ? [#56147 (review)](https://github.com/nodejs/node/pull/56147#pullrequestreview-2524315487)

Following her suggestion in the comment https://github.com/nodejs/node/pull/56147#discussion_r1898670399, I have written an example in the documentation to explain why we need this API. https://github.com/nodejs/node/pull/56147/files#diff-fc79b2d1ad702cfaf107d5880b73e8360b36273edda73f128c00641637435c3cR1368