@addaleax 2025-01-03
@theweipeng I was comparing the current version against always using UTF16-LE, to be clear, not against UTF-8.